add language choice support; improve ui; code clean-up
I added a settings panel that allows the user to select the current language to highlight. The cache was made language aware to support this.

The UI was made less intrusive by generally not displaying unnecessary things unless requested, all of which can be dismissed at any time.

Several code items were put into the .dll as resources, for easier coding. The settings panel that was added is invoked as a resource as well.

Added a command path that allows the bho to execute predefined commands on the page through javascript. See SendLwtCommand();

General code cleanup performed. Several minor warnings were addressed. Other supporting code was adding, and old code removed.